Fixed minor bug in 0.2.4 release

+++ pakrat/repo.py
@@ -81,7 +81,7 @@ def create_metadata(repo, packages=None, comps=None):
     generator.doRepoMetadata()
     generator.doFinalMove()
 
-    if os.path.exists(groupdir):
+    if comps and os.path.exists(groupdir):
         shutil.rmtree(groupdir)
 
 def sync(repo, dest, version, delete=False):
